pyproject.toml
[project]
name = "datatalk-cli"
version = "0.1.23"
description = "Ask questions about your CSV, Excel or Parquet data in natural language."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Vitaly Tsaplin", email = "vitaly@tsaplin.com"},
]
dependencies = [
    "duckdb>=0.8.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "openai>=1.102.0",
    "rich>=13.0.0",
    "openpyxl>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/vtsaplin/datatalk-cli"
Repository = "https://github.com/vtsaplin/datatalk-cli"
Issues = "https://github.com/vtsaplin/datatalk-cli/issues"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]

[project.scripts]
dtalk = "datatalk.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["datatalk*"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
